backup_parallel_20251022_145240/requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
anthropic==0.39.0
python-dotenv==1.0.0
